+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AppController } from 'src/app.controller';
 import { AppService } from 'src/app.service';
-import { UserModule } from 'src/modules/user.module';
-import { TrackModule } from 'src/modules/track.module';
+import { UserModule } from './routes/user/user.module';
+import { TrackModule } from './routes/track/track.module';
 
 @Module({
   imports: [UserModule, TrackModule],
