--- src/modules/user.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserController } from 'src/controllers/user.controller';
-import { UserService } from 'src/service/user.service';
-
-@Module({
-  controllers: [UserController],
-  providers: [UserService],
-})
-export class UserModule {}
