+++ src/routes/track/track.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { TrackController } from './track.controller';
+import TrackService from './track.service';
+
+@Module({
+  controllers: [TrackController],
+  providers: [TrackService],
+})
+export class TrackModule { }
